+++ CHANGELOG.md
@@ -46,6 +46,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix markdown formatting with backticks in API overview
 - Fix an issue that a dev dependnecy @types/dom-mediacapture-record is not getting installed
 - Fix typo in README.md file
+- Ensure that attendee presence leave and join are ordered correctly
 
 ## [1.2.1] - 2020-03-20